% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{accel}
\alias{accel}
\title{Return accelerations}
\usage{
accel(t, x, y)
}
\arguments{
\item{t}{The times vector}

\item{x}{The x positions}

\item{y}{The y positions}
}
\value{
The accelerations
}
\description{
Return accelerations
}
\seealso{
\code{\link{speed}, \link{approx_derivative}}
}
