% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_displacement}
\alias{append_displacement}
\title{Return a dataframe with information about the time-to-time displacements}
\usage{
append_displacement(data)
}
\arguments{
\item{data}{A dataframe containing t, x and y}
}
\value{
A data frame including all the dynamical information, including displacements
}
\description{
The displacement is a bit more complicated than other dynamical variables,
as it depends on the sampling frequency. If you are subsampling, always re-run
append_displacement after subsampling.
}
\seealso{
\code{\link{append_dynamics}, \link{speed}}
}
