% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_dynamics}
\alias{append_dynamics}
\title{Return a data frame with extra columns with dynamical information}
\usage{
append_dynamics(data, append.displacement = TRUE)
}
\arguments{
\item{data}{A dataframe containing t, x and y}

\item{append.displacement}{(Optional) Set it to FALSE to not calculate displacements. Useful if the data is going to be resampled}
}
\value{
A data frame including instantaneous dynamical variables, such as speed and acceleration
}
\description{
Return a data frame with extra columns with dynamical information
}
\seealso{
\code{\link{speed}, \link{accel}, \link{append_displacement}}
}
