% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{append_polar_coordinates}
\alias{append_polar_coordinates}
\title{Returns a data frame with extra columns with polar coordinates}
\usage{
append_polar_coordinates(data_loc)
}
\arguments{
\item{data_loc}{The clean data from a given location}
}
\value{
A data frame including the polar coordinates
}
\description{
Returns a data frame with extra columns with polar coordinates
}
