% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{approx_derivative}
\alias{approx_derivative}
\title{Approximate derivative}
\usage{
approx_derivative(t, x)
}
\arguments{
\item{t}{Vector of times}

\item{x}{Vector of values}
}
\value{
A vector (of the same size of t) representing the numerical derivative
}
\description{
Approximate derivative
}
\seealso{
\code{\link{speed}, \link{accel}}
}
