% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clean}
\alias{clean}
\title{Clean and filter the raw input from import}
\usage{
clean(raw)
}
\arguments{
\item{raw}{The raw data}
}
\value{
The clean data
}
\description{
Clean and filter the raw input from import
}
\seealso{
\code{\link{import}}
}
