% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{curvature}
\alias{curvature}
\title{Return curvatures}
\usage{
curvature(t, x, y)
}
\arguments{
\item{t}{The times vector}

\item{x}{The x positions}

\item{y}{The y positions}
}
\value{
The local curvature
}
\description{
Return curvatures
}
\seealso{
\code{\link{speed}, \link{accel}, \link{curvature_radius}}
}
