% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{curvature_radius}
\alias{curvature_radius}
\title{Return curvature radius}
\usage{
curvature_radius(t, x, y)
}
\arguments{
\item{t}{The times vector}

\item{x}{The x positions}

\item{y}{The y positions}
}
\value{
The local curvature radius
}
\description{
Return curvature radius
}
\seealso{
\code{\link{speed}, \link{accel}, \link{curvature}}
}
