% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{displacement}
\alias{displacement}
\title{Return displacements}
\usage{
displacement(x, y)
}
\arguments{
\item{x}{The x positions}

\item{y}{The y positions}
}
\value{
The displacements between a position and its previous
}
\description{
Return displacements
}
