% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mov}
\alias{example_mov}
\title{Example data set}
\format{
A data frame with 3000 observations of:
\describe{
  \item{x}{horizontal position}
  \item{y}{vertical position}
  \item{t}{time}
  ...
}
}
\description{
Experimental sample of 3000 positions of a macroinvertebrate
}
