% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{get_polar_coordinates}
\alias{get_polar_coordinates}
\title{Get polar coordinates}
\usage{
get_polar_coordinates(x, y, origin = c(0, 0))
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector if y coordinates}

\item{origin}{(Default = c(0, 0)) Position of the origin of coordinates}
}
\value{
Data frame with radius (r) and angle vectors (th)
}
\description{
Get polar coordinates
}
