% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{is_even}
\alias{is_even}
\title{Return true if an only if x is an even integer}
\usage{
is_even(x)
}
\arguments{
\item{x}{An integer number}
}
\value{
A boolean indicating if the input is even
}
\description{
Return true if an only if x is an even integer
}
