% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_kinematic.R
\name{plot_kinematics}
\alias{plot_kinematics}
\title{plot_shiny: The generic function for interactive plots of functional data analyses

Interactive Plotting for Kinematic Data}
\usage{
plot_kinematics(data, ...)
}
\arguments{
\item{data}{dataset to be plotted}
}
\description{
plot_shiny: The generic function for interactive plots of functional data analyses

Interactive Plotting for Kinematic Data
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

