% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{remove_unrealistic_speeds}
\alias{remove_unrealistic_speeds}
\title{Remove unrealistic speeds}
\usage{
remove_unrealistic_speeds(input_data)
}
\arguments{
\item{input_data}{A kinematic data frame}
}
\value{
A kinematic data frame, with unrealistic speeds filtered out
}
\description{
Remove unrealistic speeds
}
