% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server_kinematics.R
\name{server_kinematics}
\alias{server_kinematics}
\title{Server function for plot_kinematics}
\usage{
server_kinematics(input, output)
}
\arguments{
\item{input}{}

\item{output}{}
}
\description{
Internal function that defines the server for plot_kinematics
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

