% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{speed}
\alias{speed}
\title{Return speeds}
\usage{
speed(t, x, y)
}
\arguments{
\item{t}{The times vector}

\item{x}{The x positions}

\item{y}{The y positions}
}
\value{
The speeds
}
\description{
Return speeds
}
\seealso{
\code{\link{accel}, \link{approx_derivative}}
}
