% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_kinematics.R
\name{ui_kinematics}
\alias{ui_kinematics}
\title{UI function for plot_kinematics}
\usage{
ui_kinematics()
}
\description{
Internal function that defines the UI for plot_kinematics
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

