% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{append_displacement}
\alias{append_displacement}
\title{Returns a dataframe extra columns with information about the time-to-time displacements}
\usage{
append_displacement(data)
}
\arguments{
\item{data}{A dataframe with basic dynamics (typically the output of append_dynamics)}
}
\value{
A data frame including al the dynamical information, including displacements
}
\description{
Returns a dataframe extra columns with information about the time-to-time displacements
}
\seealso{
\code{\linl{append_dynamics}, \link{speed}, \link{accel}}
}
